mkdocs/requirements.txt
PyYAML==6.0.1
mkdocs-material==9.5.15
mkdocs-video==1.5.0
mkdocs-literate-nav==0.6.1
